requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
black==24.4.2
Brotli==1.1.0
click==8.1.7
dj-database-url==2.2.0
Django==5.0.6
django-ninja==1.1.0
gunicorn==22.0.0
h11==0.14.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
psycopg2-binary==2.9.9
pydantic==2.7.4
pydantic_core==2.18.4
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.12.2
uvicorn==0.30.1
whitenoise==6.7.0
